+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfig.kt
@@ -19,25 +19,6 @@ data class MiniAppSdkConfig(
 ) {
     internal val key = "$baseUrl-$isPreviewMode-$rasProjectId-$subscriptionKey"
 
-    @Deprecated("Use isPreviewMode instead of isTestMode, Use rasProjectId instead of rasAppId")
-    constructor(
-        baseUrl: String,
-        rasAppId: String,
-        rasProjectId: String = rasAppId,
-        subscriptionKey: String,
-        hostAppVersionId: String = "",
-        hostAppUserAgentInfo: String,
-        isTestMode: Boolean,
-        isPreviewMode: Boolean = isTestMode
-    ) : this(
-        baseUrl = baseUrl,
-        rasProjectId = rasProjectId,
-        subscriptionKey = subscriptionKey,
-        hostAppVersionId = hostAppVersionId,
-        hostAppUserAgentInfo = hostAppUserAgentInfo,
-        isPreviewMode = isPreviewMode
-    )
-
     init {
         when {
             !isBaseUrlValid(baseUrl) ->